Reorganize object detection files
